pyproject.toml
[project]
name = "judge-model-evaluation"
version = "0.1.0"
description = "LLM-as-a-judge meta-evaluation toolkit for EVOUNA."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "matplotlib>=3.9.0",
  "numpy>=2.2.0",
  "opentelemetry-api>=1.31.1",
  "opentelemetry-exporter-otlp-proto-grpc>=1.31.1",
  "opentelemetry-sdk>=1.31.1",
  "pandas>=2.2.3",
  "pyarrow>=18.1.0",
  "pydantic>=2.10.0",
  "pyyaml>=6.0.2",
  "requests>=2.32.3",
]

[project.scripts]
judge-eval = "judge_eval.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.3.4",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/judge_eval"]

[tool.pytest.ini_options]
testpaths = ["tests"]
